Added money pattern

+++ include/server/domain/weapon.hpp
@@ -14,7 +14,7 @@ public:
     using pointer = std::shared_ptr<weapon>;
 
     explicit weapon(
-        int id, std::string name, int value, int level, int attack );
+        int id, std::string name, double value, int level, int attack );
 
     [[nodiscard]] auto get_attack() const -> int;
     void set_attack( int attack );
